src/python/requirements.txt
python-can==4.5.0
netifaces2==0.0.22
paho-mqtt==2.1
colorlog==6.9.0